refactor(ppc): split PPCFunc into rex/ppc/func.h to keep SIMDe out of 81 TUs
